+++ app/uk/gov/hmrc/bindingtariffclassification/model/Paged.scala
@@ -44,7 +44,7 @@ object Paged {
   def apply[T](results: Seq[T], pagination: Pagination, resultCount: Long): Paged[T] =
     Paged(results, pagination.page, pagination.pageSize, resultCount)
 
-  def apply[T](results: Seq[T]): Paged[T] = Paged(results, Pagination(), results.size)
+  def apply[T](results: Seq[T]): Paged[T] = Paged(results, Pagination(), results.size.toLong)
 
   implicit def format[T](implicit fmt: Format[T]): Format[Paged[T]] =
     Format[Paged[T]](Reads[Paged[T]](reads), Writes[Paged[T]](writes))
@@ -59,7 +59,7 @@ object Paged {
           },
           (js \ "pageIndex").as[Int],
           (js \ "pageSize").as[Int],
-          (js \ "resultCount").as[Int]
+          (js \ "resultCount").as[Long]
         )
       ).map(JsSuccess(_))
         .recover {
